+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 
 ### Unreleased
 
+### v1.0.1 (2024-07-22)
+
+* Add readme and license
+* Add some blackbox testing
+* Throw an exception instead of silently overwriting if response handlers attempt to send more than one value
+  for the same header.
+* Improve handling if code starts output before the framework even boots.
+
+
 ### v1.0.0 (2024-07-16)
 
 * First version, extracted from host project